refactoring: the implementation of the controller interfaces has been moved to the impl folder
