+++ src/main/java/ru/gbteam/lms/controller/AdminController.java
@@ -1,4 +1,4 @@
-package ru.gbteam.lms.controller.impl;
+package ru.gbteam.lms.controller;
 
 import org.springframework.security.access.annotation.Secured;
 import org.springframework.ui.Model;
