+++ src/main/java/ru/gbteam/lms/controller/ExceptionHandlerController.java
@@ -1,4 +1,4 @@
-package ru.gbteam.lms.controller.impl;
+package ru.gbteam.lms.controller;
 
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.servlet.ModelAndView;
