+++ src/main/java/ru/gbteam/lms/controller/IndexController.java
@@ -1,4 +1,4 @@
-package ru.gbteam.lms.controller.impl;
+package ru.gbteam.lms.controller;
 
 import org.springframework.web.bind.annotation.GetMapping;
