+++ src/main/java/ru/gbteam/lms/controller/impl/AdminControllerImpl.java
@@ -1,11 +1,10 @@
-package ru.gbteam.lms.controller;
+package ru.gbteam.lms.controller.impl;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.security.access.annotation.Secured;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
-import ru.gbteam.lms.controller.impl.AdminController;
+import ru.gbteam.lms.controller.AdminController;
 import ru.gbteam.lms.dto.UserDto;
 import ru.gbteam.lms.exception.NotFoundException;
 import ru.gbteam.lms.model.Role;
