+++ src/main/java/ru/gbteam/lms/controller/impl/IndexControllerImpl.java
@@ -1,4 +1,4 @@
-package ru.gbteam.lms.controller;
+package ru.gbteam.lms.controller.impl;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
