+++ src/test/java/ru/gbteam/lms/controller/CourseControllerTest.java
@@ -11,6 +11,7 @@ import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilde
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.test.web.servlet.result.MockMvcResultMatchers;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import ru.gbteam.lms.controller.impl.CourseControllerImpl;
 import ru.gbteam.lms.model.Course;
 import ru.gbteam.lms.model.User;
 import ru.gbteam.lms.service.CourseServiceFacade;
