+++ src/test/java/ru/gbteam/lms/controller/ExceptionHandlerControllerTest.java
@@ -6,6 +6,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.servlet.ModelAndView;
+import ru.gbteam.lms.controller.impl.ExceptionHandlerControllerImpl;
 import ru.gbteam.lms.exception.NotFoundException;
 
 class ExceptionHandlerControllerTest {
